migrateMeetingTime.ts usages changed, file name changed, test is moved into src.
