+++ matrix-meetings-bot/src/client/MeetingClient.ts
@@ -53,7 +53,7 @@ import {
   getMeetingStartTime,
 } from '../shared';
 import { extractOxRrule } from '../util/extractOxRrule';
-import { migrateMeetingTime } from '../util/IMeetingTime';
+import { migrateMeetingTime } from '../util/migrateMeetingTime';
 import { templateHelper } from '../util/TemplateHelper';
 
 @Injectable()
@@ -238,25 +238,19 @@ export class MeetingClient {
       params
     );
 
-    const meetingTime = migrateMeetingTime(
-      {
-        start_time: meetingCreate.start_time,
-        end_time: meetingCreate.end_time,
-        calendar: meetingCreate.calendar,
-      },
+    // change data model if meeting is OX with non-empty rrules
+    const { start_time, end_time, calendar } = migrateMeetingTime(
+      meetingCreate,
       extractOxRrule(meetingCreate)
     );
 
     const meetingsMetadataEventContent: IMeetingsMetadataEventContent = {
-      start_time: meetingTime.start_time,
-      end_time: meetingTime.end_time,
-      calendar: meetingTime.calendar,
-      force_deletion_at: getForceDeletionTime(
-        autoDeletionOffset,
-        meetingTime.calendar
-      ),
+      start_time,
+      end_time,
+      calendar,
+      force_deletion_at: getForceDeletionTime(autoDeletionOffset, calendar),
       auto_deletion_offset:
-        meetingTime.calendar === undefined ? autoDeletionOffset : undefined,
+        calendar === undefined ? autoDeletionOffset : undefined,
       creator: userContext.userId,
       external_data: meetingCreate.external_data,
     };
